delete redundent files
