--- duke.txt
@@ -1,1 +1,0 @@
-todo [T][✘] read book
