--- src/main/java/Level0.java
@@ -1,13 +1,0 @@
-public class Level0 {
-
-        public static void main(String[] args){
-
-            System.out.println("____________________________________________________________\n");
-            System.out.println(" Hello! I'm Duke\n");
-            System.out.println(" What can I do for you?\n");
-            System.out.println("____________________________________________________________\n");
-            System.out.println(" Bye. Hope to see you again soon!\n");
-            System.out.println("____________________________________________________________\n");
-        }
-    }
-
