+++ src/main/java/duke.txt
@@ -1,0 +1,1 @@
+todo [T][?] read book
